requirements.txt
openai==1.55.3
pydantic>=2.0,<3.0
PyYAML>=6.0,<7.0
pypdf==4.3.1
